+++ objc2-foundation/CHANGELOG.md
@@ -28,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added `NSPoint`.
 * Added `NSSize`.
 * Added `NSRect`.
+* Implement `Borrow` and `BorrowMut` for all objects.
 
 ### Changed
 * **BREAKING**: Removed the following helper traits in favor of inherent